feat(Dimensions): Update dimensions when window size / orientation changes
Adds feature to update the dimensions when the window size changes.
